+++ CHANGELOG.md
@@ -58,6 +58,7 @@
 
 ### Core Bug Fixes / Improvements
 
+* VALIDATOR: Improvements to CPU validation system to support initial generation of V20 tests.
 * 8088: Refactored instruction decode to a table-based lookup, replaced custom flags with values from group decode ROM.
 * 8088: Converted SegmentOverride enum to Option<Segment>, simplifying segment resolution logic in many places
 * 8088: New 7-cycle bus access logic